XmppDateTime: accept `+0000` as a timezone string
There was a bug in smack3 where <delay> timestamps were generated with
a timezone string of `+0000` instead of `+00:00` or `Z`.  When
processing these <delay> elements, the current parser will crash with
an "Unparseable date" exception, either bringing down the connection or
at least failing to process the <delay/> element.

This patch allows processing of both `+00:00` and `+0000` timezone
strings and correctly interpreting the time from either. While this is
not in conformance to XEP-0082, there is no damage in being liberal in
what we accept.

Signed-off-by: Georg Lukas <georg@op-co.de>